updated rust templates
