+++ checkout/rust/cart-transform/bundles/Cargo.toml.liquid
@@ -8,7 +8,7 @@ rust-version = "1.62"
 serde = { version = "1.0.13", features = ["derive"] }
 serde_with = "1.13.0"
 serde_json = "1.0"
-shopify_function = "0.3.0"
+shopify_function = "0.4.0"
 graphql_client = "0.13.0"
 
 [profile.release]
