+++ checkout/rust/cart-transform/default/.gitignore
@@ -1,3 +1,3 @@
 /target
 Cargo.lock
-.output.graphql
+*.output.graphql
