+++ checkout/rust/cart-transform/default/Cargo.toml.liquid
@@ -1,5 +1,5 @@
 [package]
-name = "cart_transform"
+name = "{{handle | replace: " ", "-" | downcase}}"
 version = "1.0.0"
 edition = "2021"
 rust-version = "1.62"
@@ -8,7 +8,7 @@ rust-version = "1.62"
 serde = { version = "1.0.13", features = ["derive"] }
 serde_with = "1.13.0"
 serde_json = "1.0"
-shopify_function = "0.3.0"
+shopify_function = "0.4.0"
 graphql_client = "0.13.0"
 
 [profile.release]
